updated 60fps code

+++ src/code/z_lifemeter.cpp
@@ -286,7 +286,7 @@ void HealthMeter_Draw(GlobalContext* globalCtx) {
     s16 totalHeartCount = gSaveContext.healthCapacity / 0x10;
     s16 fullHeartCount = gSaveContext.health / 0x10;
     s32 pad2;
-    f32 beatingHeartPulsingSize = interfaceCtx->beatingHeartOscillator.toFloat() * 0.1f;
+    f32 beatingHeartPulsingSize = interfaceCtx->beatingHeartOscillator * 0.1f;
     s32 curCombineModeSet = 0;
     u8* curBgImgLoaded = NULL;
     s32 ddHeartCountMinusOne = gSaveContext.inventory.defenseHearts - 1;
